Wednesday, June 14, 2023 at 8:45 PM Central Daylight Time

+++ commands/cogs/Server/WadderVerification.py
@@ -2,8 +2,15 @@ import nextcord
 import json
 from nextcord.ext import commands
 from nextcord.ui import Button, View
+from pymongo import MongoClient
+import urllib.parse
 
-VERIFICATION_DATA_FILE = 'data/verification_data.json'
+# MongoDB connection details
+username = urllib.parse.quote_plus("apwade75009")
+password = urllib.parse.quote_plus("Celina@12")
+cluster = MongoClient(f"mongodb+srv://{username}:{password}@quantaai.irlbjcw.mongodb.net/")
+db = cluster["QuantaAI"]
+verification_collection = db["verification"]
 
 # ! This is not ready for production
 
@@ -25,18 +32,13 @@ class VerificationView(View):
 class VerificationCog(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
-        self.verification_data = self.load_verification_data()
+        self.verification_collection = verification_collection
 
-    def load_verification_data(self):
-        try:
-            with open(VERIFICATION_DATA_FILE, 'r') as f:
-                return json.load(f)
-        except FileNotFoundError:
-            return {}
+    def get_verification_data(self, guild_id):
+        return self.verification_collection.find_one({"guild_id": guild_id})
 
-    def save_verification_data(self):
-        with open(VERIFICATION_DATA_FILE, 'w') as f:
-            json.dump(self.verification_data, f, indent=4)
+    def save_verification_data(self, verification_data):
+        self.verification_collection.replace_one({"guild_id": verification_data["guild_id"]}, verification_data, upsert=True)
 
     async def is_admin(self, user: nextcord.User):
         return user.guild_permissions.administrator
@@ -51,32 +53,31 @@ class VerificationCog(commands.Cog):
         return channel
 
     async def find_verification_channels(self):
-        for guild_id, guild_data in self.verification_data.items():
-            guild = self.bot.get_guild(int(guild_id))
-            if guild:
-                channel_id = guild_data.get('channel_id')
-                if channel_id:
-                    channel = guild.get_channel(int(channel_id))
-                    if channel:
-                        # Delete all previous verification messages in the channel
-                        async for message in channel.history(limit=100):
-                            if message.author == self.bot.user and message.content.startswith("Click the button below"):
-                                await message.delete()
+        for guild in self.bot.guilds:
+            verification_data = self.get_verification_data(guild.id)
+            if verification_data and verification_data.get('channel_id'):
+                channel_id = verification_data['channel_id']
+                channel = guild.get_channel(channel_id)
+                if channel:
+                    # Delete all previous verification messages in the channel
+                    async for message in channel.history(limit=100):
+                        if message.author == self.bot.user and message.content.startswith("Click the button below"):
+                            await message.delete()
 
-                        # Send a new verification message with a button
-                        view = VerificationView(self)
-                        await channel.send("Click the button below to get verified:", view=view)
+                    # Send a new verification message with a button
+                    view = VerificationView(self)
+                    await channel.send("Click the button below to get verified:", view=view)
 
     async def verify_member(self, interaction: nextcord.Interaction):
         await interaction.response.defer()
         member = interaction.user
         guild = interaction.guild
-        guild_data = self.verification_data.get(str(guild.id), {})
-        role_id = guild_data.get('role_id')
+        verification_data = self.get_verification_data(guild.id)
 
-        if role_id:
+        if verification_data and verification_data.get('role_id'):
+            role_id = verification_data['role_id']
             verified_role = guild.get_role(role_id)
-            if verified_role not in member.roles:
+            if verified_role and verified_role not in member.roles:
                 await member.add_roles(verified_role)
                 dm_channel = await member.create_dm()
                 await dm_channel.send("You've been successfully verified! You can now access the server channels.")
@@ -89,16 +90,17 @@ class VerificationCog(commands.Cog):
     @commands.command(name="svrole")
     async def set_verification_role(self, ctx: commands.Context, role_id: int):
         if await self.is_admin(ctx.author):
-            if str(ctx.guild.id) not in self.verification_data:
-                self.verification_data[str(ctx.guild.id)] = {}
+            verification_data = self.get_verification_data(ctx.guild.id)
+            if not verification_data:
+                verification_data = {"guild_id": ctx.guild.id}
 
-            self.verification_data[str(ctx.guild.id)]['role_id'] = role_id
+            verification_data['role_id'] = role_id
 
-            if 'channel_id' not in self.verification_data[str(ctx.guild.id)]:
+            if 'channel_id' not in verification_data:
                 channel = await self.create_verification_channel(ctx.guild)
-                self.verification_data[str(ctx.guild.id)]['channel_id'] = channel.id
+                verification_data['channel_id'] = channel.id
 
-            self.save_verification_data()
+            self.save_verification_data(verification_data)
             await ctx.send(f"Verification role successfully set to role ID {role_id}. A verification channel with a button has been created.")
         else:
             await ctx.send("You don't have permission to use this command.")
